requirements/production.txt
-r base.txt

gunicorn==20.1.0
sentry-sdk==1.1.0

gevent==22.10.2

django-anymail[mailgun]